refactor(ts): Make deep file structures (#36)
Separate files by their roles, and merge tests.
