+++ deno.json
@@ -1,7 +1,6 @@
 {
   "$schema": "https://deno.land/x/deno/cli/schemas/config-file.v1.json",
   "fmt": { "exclude": ["LICENSE", ".github/**/*.md"] },
-  "test": { "include": ["src/", "test/"] },
   "tasks": {
     "run": "deno run --env='.env' --allow-env='REPOSITORY_OWNER,REPOSITORY_NAME,REPOSITORY_PATH,GITHUB_TOKEN' --allow-net='0.0.0.0,api.github.com'",
     "start": "deno task run ./src/server.ts",
@@ -10,15 +9,13 @@
     "cov": "deno task test --coverage && deno coverage --lcov > coverage.lcov"
   },
   "imports": {
-    "@5ouma/reproxy/libs": "./src/libs/mod.ts",
-    "@5ouma/reproxy/types": "./src/types/mod.ts",
     "@oak/oak": "jsr:@oak/oak@16.1.0",
     "@octokit/rest": "https://esm.sh/@octokit/rest@21.0.1",
     "@std/assert": "jsr:@std/assert@1.0.2",
     "@std/fmt": "jsr:@std/fmt@1.0.0",
     "@std/http": "jsr:@std/http@1.0.2",
     "@std/http/user-agent": "jsr:@std/http@~0.223/user-agent",
-    "@std/url": "jsr:@std/url@0.225.0"
+    "@std/path": "jsr:@std/path@1.0.2"
   },
   "deploy": {
     "project": "reproxy",
