+++ src/libs/middlewares/redirect.ts
@@ -1,8 +1,9 @@
 import type { RouterContext } from "@oak/oak";
 import { STATUS_CODE } from "@std/http/status";
 import { UserAgent } from "@std/http/user-agent";
-import { join } from "@std/url";
-import { getRepository } from "./env.ts";
+import { join } from "@std/path";
+
+import { getRepository } from "../utils/env.ts";
 
 export function redirect<R extends string>(
   ctx: RouterContext<R>,
@@ -10,13 +11,15 @@ export function redirect<R extends string>(
   ref: string = "master",
 ): boolean {
   const repository = getRepository();
-  const url = join(
-    new URL("https://github.com"),
-    repository.owner,
-    repository.name,
-    "blob",
-    ref,
-    repository.path,
+  const url = new URL(
+    join(
+      repository.owner,
+      repository.name,
+      "blob",
+      ref,
+      repository.path,
+    ),
+    "https://github.com",
   );
 
   if (!userAgent?.browser.name) return false;
