+++ src/libs/utils/env_test.ts
@@ -1,6 +1,7 @@
 import { assertEquals } from "@std/assert";
-import { getRepository } from "../../src/libs/env.ts";
-import { clearRepo, exportRepo, testRepo } from "../utils.ts";
+
+import { getRepository } from "./env.ts";
+import { clearRepo, exportRepo, testRepo } from "../test_utils.ts";
 
 Deno.test("Get Repository Env", async (t: Deno.TestContext) => {
   await t.step("Normal", () => {
